+++ archives-manage-entity/src/main/java/com/yintu/rixing/system/SysApprovalProcessConfiguration.java
@@ -38,7 +38,7 @@ public class SysApprovalProcessConfiguration extends BaseEntity {
 
     @ApiModelProperty(value = "顺序")
     @TableField("`order`")
-    private Short order;
+    private Integer order;
 
 
 }
